half merge cp_r and cp_filtered

+++ src/bootstrap/lib.rs
@@ -1538,18 +1538,7 @@ impl Build {
         if self.config.dry_run() {
             return;
         }
-        for f in self.read_dir(src) {
-            let path = f.path();
-            let name = path.file_name().unwrap();
-            let dst = dst.join(name);
-            if t!(f.file_type()).is_dir() {
-                t!(fs::create_dir_all(&dst));
-                self.cp_r(&path, &dst);
-            } else {
-                let _ = fs::remove_file(&dst);
-                self.copy(&path, &dst);
-            }
-        }
+        self.recurse_(src, dst, Path::new(""), &|_| true, false)
     }
 
     /// Copies the `src` directory recursively to `dst`. Both are assumed to exist
@@ -1557,11 +1546,20 @@ impl Build {
     /// by returning `false` from the filter function.
     pub fn cp_filtered(&self, src: &Path, dst: &Path, filter: &dyn Fn(&Path) -> bool) {
         // Immediately recurse with an empty relative path
-        self.recurse_(src, dst, Path::new(""), filter)
+        self.recurse_(src, dst, Path::new(""), filter, true)
     }
 
     // Inner function does the actual work
-    fn recurse_(&self, src: &Path, dst: &Path, relative: &Path, filter: &dyn Fn(&Path) -> bool) {
+    //
+    // FIXME: consider merging cp_filtered and cp_r into one function
+    fn recurse_(
+        &self,
+        src: &Path,
+        dst: &Path,
+        relative: &Path,
+        filter: &dyn Fn(&Path) -> bool,
+        remove_dst_dir: bool,
+    ) {
         for f in self.read_dir(src) {
             let path = f.path();
             let name = path.file_name().unwrap();
@@ -1570,9 +1568,11 @@ impl Build {
             // Only copy file or directory if the filter function returns true
             if filter(&relative) {
                 if t!(f.file_type()).is_dir() {
-                    let _ = fs::remove_dir_all(&dst);
+                    if remove_dst_dir {
+                        let _ = fs::remove_dir_all(&dst);
+                    }
                     self.create_dir(&dst);
-                    self.recurse_(&path, &dst, &relative, filter);
+                    self.recurse_(&path, &dst, &relative, filter, remove_dst_dir);
                 } else {
                     let _ = fs::remove_file(&dst);
                     self.copy(&path, &dst);
